refactor(dts): implement cjsReexport stub via Rolldown plugin
