+++ src/build.ts
@@ -1,5 +1,3 @@
-import { writeFile } from 'node:fs/promises'
-import path from 'node:path'
 import { bold, green } from 'ansis'
 import { clearRequireCache } from 'import-without-cache'
 import {
@@ -311,10 +309,6 @@ async function buildSingle(
     await copy(config)
     await buildExe(config, chunks)
 
-    if (format === 'cjs' && dts && dts.cjsReexport && isDualFormat) {
-      await writeCjsDtsReexports(chunks, outDir, config.write !== false)
-    }
-
     if (!hasBuilt) {
       await done(bundle)
     }
@@ -324,29 +318,5 @@ async function buildSingle(
 
     ab?.abort()
     ab = executeOnSuccess(config)
-  }
-}
-
-async function writeCjsDtsReexports(
-  chunks: RolldownChunk[],
-  outDir: string,
-  write: boolean,
-): Promise<void> {
-  for (const chunk of chunks) {
-    if (chunk.type !== 'chunk') continue
-
-    // Match CJS JS output files: .cjs (fixed extension) or .js (non-fixed)
-    const match = chunk.fileName.match(/^(.*)\.(cjs|js)$/)
-    if (!match) continue
-
-    const baseName = match[1]
-    const dCtsName =
-      match[2] === 'cjs' ? `${baseName}.d.cts` : `${baseName}.d.ts`
-    const dMtsBasename = path.basename(`${baseName}.d.mts`)
-    const content = `export * from './${dMtsBasename}'\n`
-
-    if (write) {
-      await writeFile(path.join(outDir, dCtsName), content)
-    }
   }
 }
